Make examples testable
Currently there are no assertions, but it is still useful to
be able to quickly generate the plots for all examples at once.

Tests are now also enabled through Gitlab CI. Note that I have
installed NumPy and Matplotlib system-wide through apt, so it
does not need to be included in the CI configuration with the
way that the server is currently configured. This may change.