Add loihi command line
The main function of the command line interface is to do all
of the commands needed to compile nengo_loihi, sync and upload
files, run a model using SLURM, and download generated plots
in one command.

The README has been updated with the new workflow, and the
interface has documentation accessible with the --help flag.

The command line can be extended in the future to do more things
like running models on ABR controlled boards.