Test examples with pytest
And test their style with flake8 and pylint during CI.

In order to be able to run the examples within the test suite
and normally from the command line, the plotting parts were
moved to an `if __name__ == "__main__"`. In the test suite,
most of those plots are still produced, but print statements
are removed, and assertions are added.

- Probing inputs with `max_time` is broken on Loihi, so
  node_ens_ens now uses `precompute=True`.
  Having `precompute=True` results in more accurate values
  for measuring input nodes, which means that loihi is actually
  farther from the reference value. This required looser tolerances
  comparing `u` and `a` in test_node_ens_ens.