Simplify test_simulator_noise
I noticed that this test was spitting out a PDF file in the
current directory. Changed it to use the pytest fixtures better,
including the `request` fixture which allows us to look at the
value of the `--target` option.

One slightly hacky change that I made was to change the
simulators in `nengo.conftest.TestConfig`. Those sims are used
for the default directory used for the `plt` fixture when the
`Simulator` fixture is not specified. By changing them, all
plots will be put in the `nengo_loihi.simulator.plots` directory
when running the test suite.