Add allclose fixture
The idea behind this is to get a general sense of how a change
affects accuracy on all models. We already do several tests
doing np.allclose on some simulated value and the desired target.
The `allclose` fixture is a drop-in replacement for `np.allclose`
that (at the moment) also records the RMSE between the simulated
and desired values. After all tests have ran, the RMSE values
recorded on successful tests are reported as the mean +/- std.
The data is there if we want to do something more sophisticated
like plot the RMSE distribution, or something of that nature.