chdir to NxSDK earlier
This seems like it will be necessary in later versions on NxSDK,
as spike generators will be switching to using snips.
Also, this chdirs back to the original working directory after
the Loihi sim is closed. It might be possible to switch back
earlier, but this seems the most conservative estimate given
what we know about how NxSDK works.