Implement multidimensional learning
The actual communication protocal is not as efficient as
it could be, but this work properly.

The choice of the function computed across the learned
connection prior to learning turns out to have a huge
effect on the behavior of the network. If the weights are
initially much smaller than they will be post-learning,
the weights can easily over/underflow. Improving the
weight discretization, however, is left for future work.

This commit also removes some writes that used to be needed
to close the snip successfully, but they appear to be
unnecessary now.