Interneuron default total rate is 1/dt
`inter_rate` is now chosen by default based on `dt`, so that the
sum of the max rates of all interneurons adds to the max firing rate
of 1/dt. Since this sum of max rates is used for the interneurons
from nodes in `splitter.py`, this ensures that models can run with
any `dt` without needing to change `inter_rate`.