Match learning between Nengo/emulator/chip
This ensures that both the emulator and chip match Nengo in terms of
basic PES learning. The emulator is now much more accurate in terms
of how the chip does learning, which allows us to account for this
when mapping the PES learning parameters (i.e., learning rate,
synapse decay) to the chip.