Refactor Simulator.run_steps
One major changes is to move the host-to-chip and chip-to-host
logic to CxSimulator and LoihiSimulator. This keeps all of the
manipulation of CxSimulator- and LoihiSimulator-internal values
internal to those classes.

The other major change is to handle each possible permutation of
`run_steps` logic independently and cache the resulting sequence
of step functions. Part of this is for readability (it is now
more clear what happens in what cases, and is split up such
that we no longer get the cyclomatic complexity warning),
but also for speed as in some cases `run_steps` could be
called within a loop.