Refactor interneurons, now called decode neurons
The biggest change is to be able to customize how many decode
neurons are used and how they are organized.
Additionally, the name "interneurons" was changed to "decode
neurons" to avoid confusion with the concept of inhibitory
interneurons in neuroscience.

This commit also removes sharing encoders between node inputs
and ensemble inputs. This greatly simplifies things, because these
no longer have to be identical.

Several decode neuron implementations are provided:
- OnOffDecodeNeurons uses identical pairs of on/off neurons for
  each dimension. This was the default behavior prior to this commit,
  and is still used to generate input spikes (for nodes).
- NoisyDecodeNeurons uses pairs of on/off neurons for each dimension
  with injected noise so that each on/off pair is different.
- Preset5DecodeNeurons uses five heterogeneous on/off pairs with
  pre-set values for each dimension.
- Preset10DecodeNeurons uses ten heterogeneous on/off pairs with
  pre-set values for each dimension. This is now the default behavior
  for decoded connections between two ensembles.

Partially addresses #129.